add a unitary test for filterInscription close test
