+++ frontend/src/components/Profil/Profil.jsx
@@ -2,6 +2,7 @@
 import { useEffect, useState } from 'react';
 import axios from 'axios';
 import SmallCards from '@components/SmallCards/SmallCards';
+import { filterInscriptionsFermées } from '@/services/filterInscriptionFermée';
 
 export default function Profil() {
     const [tournaments, setTournaments] = useState([]);
@@ -44,18 +45,18 @@ export default function Profil() {
     };
 
     // Fonction pour filtrer les tournois avec inscriptions fermées
-    const filterInscriptionsFermées = () => {
-        const today = new Date();
-        const year = today.getFullYear();
-        const month = String(today.getMonth() + 1).padStart(2, '0');
-        const day = String(today.getDate()).padStart(2, '0');
-        const formattedDate = `${year}-${month}-${day}`;
+    // const filterInscriptionsFermées = () => {
+    //     const today = new Date();
+    //     const year = today.getFullYear();
+    //     const month = String(today.getMonth() + 1).padStart(2, '0');
+    //     const day = String(today.getDate()).padStart(2, '0');
+    //     const formattedDate = `${year}-${month}-${day}`;
 
-        return tournaments.filter(tournament => {
-            const participants = participantsData[tournament.id] || [];
-            return !(tournament.total_players > participants.length && tournament.date > formattedDate);
-        });
-    };
+    //     return tournaments.filter(tournament => {
+    //         const participants = participantsData[tournament.id] || [];
+    //         return !(tournament.total_players > participants.length && tournament.date > formattedDate);
+    //     });
+    // };
 
     return (
         <div>
@@ -101,7 +102,7 @@ export default function Profil() {
             {/* Affichage des anciens tournois (inscriptions fermées) */}
             <div className="flex flex-col items-center p-4">
                 <div className="flex flex-wrap justify-center w-full gap-4">
-                    {filterInscriptionsFermées().map((tournament, index) => (
+                    {filterInscriptionsFermées(tournaments, participantsData).map((tournament, index) => (
                         <SmallCards key={tournament.id} tournament={tournament} index={index} />
                     ))}
                 </div>
